Rename toolbar controls
